redone transplant simulations and I&S anova analysis. -- LJS

+++ make_all_movie_hidden.m
@@ -2,15 +2,19 @@
 % saves to /avi_mat
 
 % close all
-  
+
 movieFig = figure('units','points','outerposition',[0 0 1600 430],'position',[1 1 1599 363],...
     'PaperPositionMode','manual','PaperUnits','points','PaperPosition',[0 0 1600 430],'visible','off');
 filePath = ['avi_mat/allmovie/',saveInfo,'/'];
 mkdir(filePath)
 frameRate = 30;    % frames per second - fewer frames will make the movie slower
 skip = round(5/60/param.tstep); % skip more time steps for faster saving
-frameCtr = 1;
-% % set(gcf,'Renderer','zbuffer')
+
+if exist('startAtFrame','var')
+    frameCtr = startAtFrame;
+else
+    frameCtr = 1;
+end
 
 caCmap = load('cmap_blue2cyan.txt');
 
@@ -31,20 +35,27 @@ for timeCtr=1:skip:numTsteps
     T = get(gca,'Position');
     set(gca,'position',[T_t(1)+0.04 T_t(2) 1-T_t(1)-T_t(3)-0.1 1-T_t(2)-T_t(4)]);
     
-%     axis image
+    %     axis image
     print(movieFig,'-r0','-dtiff',[filePath,int2str(frameCtr),'.tiff'])
     frameCtr = frameCtr + 1;
+    if exist('stopTime','var')&&t_save(timeCtr)>=stopTime
+        break
+    end
 end
 close(movieFig)
 %% using external video compiling
-disp('compiling the movie now');
-
-%help on this to be found here:
-%https://libav.org/avconv.html
-
-system(['avconv -i ', filePath, '%d.tiff -r ', int2str(frameRate), ' -qscale 10 -y ',filePath,saveInfo,'.mp4']);
-% To have a constant quality (but a variable bitrate), use the option 
-% ?-qscale n? when ?n? is between 1 (excellent quality) and 31 (worst quality). 
-% delete the individual frames
-system(['rm -f ' filePath '*.tiff']);
+if exist('keepFrames','var')&&keepFrames==1
+    disp('keeping individual franes')
+else
+    disp('compiling the movie now');
+    %help on this to be found here:
+    %https://libav.org/avconv.html
+    
+    system(['avconv -i ', filePath, '%d.tiff -r ', int2str(frameRate), ' -qscale 10 -y ',filePath,saveInfo,'.mp4']);
+    % To have a constant quality (but a variable bitrate), use the option
+    % ?-qscale n? when ?n? is between 1 (excellent quality) and 31 (worst quality).
+    
+    % delete the individual frames
+    system(['rm -f ' filePath '*.tiff']);
+end
 disp(['saved to ',filePath])
